Move OsmDroidMapConfigurator to osmdroid
