+++ collect_app/src/main/java/org/odk/collect/android/geo/MapConfiguratorProvider.java
@@ -22,7 +22,8 @@ import com.mapbox.mapboxsdk.maps.Style;
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.geo.GoogleMapConfigurator.GoogleMapTypeOption;
-import org.odk.collect.android.geo.OsmDroidMapConfigurator.WmsOption;
+import org.odk.collect.osmdroid.OsmDroidMapConfigurator;
+import org.odk.collect.osmdroid.OsmDroidMapConfigurator.WmsOption;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.maps.MapConfigurator;
 import org.odk.collect.osmdroid.WebMapService;
