+++ osmdroid/build.gradle
@@ -46,6 +46,7 @@ dependencies {
 
     implementation Dependencies.osmdroid
     implementation Dependencies.androidx_fragment_ktx
+    implementation Dependencies.androidx_preference_ktx
     implementation Dependencies.android_material
     implementation Dependencies.timber
     implementation Dependencies.play_services_location
