+++ osmdroid/src/main/java/org/odk/collect/osmdroid/OsmDroidMapConfigurator.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.geo;
+package org.odk.collect.osmdroid;
 
 import static org.odk.collect.settings.keys.ProjectKeys.KEY_REFERENCE_LAYER;
 import static kotlin.collections.SetsKt.setOf;
@@ -8,12 +8,9 @@ import android.os.Bundle;
 
 import androidx.preference.Preference;
 
-import org.odk.collect.android.R;
 import org.odk.collect.androidshared.ui.PrefUtils;
 import org.odk.collect.maps.MapConfigurator;
 import org.odk.collect.maps.layers.MbtilesFile;
-import org.odk.collect.osmdroid.OsmDroidMapFragment;
-import org.odk.collect.osmdroid.WebMapService;
 import org.odk.collect.shared.settings.Settings;
 
 import java.io.File;
@@ -21,13 +18,13 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-class OsmDroidMapConfigurator implements MapConfigurator {
+public class OsmDroidMapConfigurator implements MapConfigurator {
     private final String prefKey;
     private final int sourceLabelId;
     private final WmsOption[] options;
 
     /** Constructs a configurator that renders just one Web Map Service. */
-    OsmDroidMapConfigurator(WebMapService service) {
+    public OsmDroidMapConfigurator(WebMapService service) {
         prefKey = "";
         sourceLabelId = 0;
         options = new WmsOption[] {new WmsOption("", 0, service)};
@@ -37,7 +34,7 @@ class OsmDroidMapConfigurator implements MapConfigurator {
      * Constructs a configurator that offers a few Web Map Services to choose from.
      * The choice of which Web Map Service will be stored in a string preference.
      */
-    OsmDroidMapConfigurator(String prefKey, int sourceLabelId, WmsOption... options) {
+    public OsmDroidMapConfigurator(String prefKey, int sourceLabelId, WmsOption... options) {
         this.prefKey = prefKey;
         this.sourceLabelId = sourceLabelId;
         this.options = options;
@@ -98,12 +95,12 @@ class OsmDroidMapConfigurator implements MapConfigurator {
         return name != null ? name : file.getName();
     }
 
-    static class WmsOption {
+    public static class WmsOption {
         final String id;
         final int labelId;
         final WebMapService service;
 
-        WmsOption(String id, int labelId, WebMapService service) {
+        public WmsOption(String id, int labelId, WebMapService service) {
             this.id = id;
             this.labelId = labelId;
             this.service = service;
